allow adding html tags + small refactoring
