+++ packages/features/ee/workflows/components/v2/TextEditor/TextEditor.tsx
@@ -6,7 +6,7 @@ import { EditorProps } from "react-draft-wysiwyg";
 import { RichUtils } from 'draft-js';
 
 import "./styles.css";
-import { AddVariablesDropdown } from "../../components/v2/AddVariablesDropdown";
+import { AddVariablesDropdown } from "../AddVariablesDropdown";
 import { FormValues } from "ee/workflows/pages/v2/workflow";
 import draftToHtml from 'draftjs-to-html';
 
@@ -29,7 +29,6 @@ function TextEditor(props: TextEditorProps) {
 
   const [editorState, setEditorState] = useState(EditorState.createWithContent(contentState));
 
-
   return (
     <div>
       <Editor
@@ -37,7 +36,7 @@ function TextEditor(props: TextEditorProps) {
         wrapperClassName="border border-gray-300 rounded-md"
         editorClassName=" p-3 -mt-3 bg-white rounded-md text-sm"
         onEditorStateChange={setEditorState}
-        onChange={() => {props.form.setValue(`steps.${props.stepNumber - 1}.reminderBody`, draftToHtml(convertToRaw(editorState.getCurrentContent())))} }
+        onChange={() => {props.form.setValue(`steps.${props.stepNumber - 1}.reminderBody`, draftToHtml(convertToRaw(editorState.getCurrentContent())).replaceAll("&lt;","<").replaceAll("&gt;", ">"))} }
         toolbarCustomButtons={[<CustomOption editorState={editorState} setEditorState={setEditorState} addVariable={props.addVariable} />]}
         toolbar={{
           options: ["inline", "fontSize", "list"],
